int -> long bc c99
according to c99:
int is at least 16 bits
long is at least 32 bits
we need 32 bits here

+++ base64.c
@@ -76,7 +76,7 @@ static const char PADDING = '=';
 
 // encode full chunk
 int encode_chunk_full(char out[static 4], const char in[static 3]) {
-    int tmp = (int)in[0] << 16 | (int)in[1] << 8 | (int)in[2];
+    long tmp = (long)in[0] << 16 | (long)in[1] << 8 | (long)in[2];
     out[0] = ALPHABET[tmp >> 18 & 63];
     out[1] = ALPHABET[tmp >> 12 & 63];
     out[2] = ALPHABET[tmp >> 6 & 63];
@@ -89,7 +89,7 @@ int encode_chunk_full(char out[static 4], const char in[static 3]) {
 int encode_chunk_partial(char out[static 4], const char in[], int length) {
     switch (length) {
         case 1: {
-            int tmp = (int)in[0] << 16;
+            long tmp = (long)in[0] << 16;
             out[0] = ALPHABET[tmp >> 18 & 63];
             out[1] = ALPHABET[tmp >> 12 & 63];
             out[2] = PADDING;
@@ -97,7 +97,7 @@ int encode_chunk_partial(char out[static 4], const char in[], int length) {
             return 2;
         }
         case 2: {
-            int tmp = (int)in[0] << 16 | (int)in[1] << 8;
+            long tmp = (long)in[0] << 16 | (long)in[1] << 8;
             out[0] = ALPHABET[tmp >> 18 & 63];
             out[1] = ALPHABET[tmp >> 12 & 63];
             out[2] = ALPHABET[tmp >> 6 & 63];
@@ -122,7 +122,7 @@ int encode_chunk(char out[static 4], const char in[], int length) {
 }
 
 int decode_chunk_full(char out[static 3], const char in[static 4]) {
-    int tmp = REVERSE_ALPHABET[(int)in[0]] << 18
+    long tmp = REVERSE_ALPHABET[(int)in[0]] << 18
             | REVERSE_ALPHABET[(int)in[1]] << 12
             | REVERSE_ALPHABET[(int)in[2]] << 6
             | REVERSE_ALPHABET[(int)in[3]];
@@ -142,13 +142,13 @@ int decode_chunk_partial(char out[], const char in[], int length) {
             return 1;
         }
         case 2: {
-            int tmp = REVERSE_ALPHABET[(int)in[0]] << 18
+            long tmp = REVERSE_ALPHABET[(int)in[0]] << 18
                     | REVERSE_ALPHABET[(int)in[1]] << 12;
             out[0] = tmp >> 16 & 255;
             return 1;
         }
         case 3: {
-            int tmp = REVERSE_ALPHABET[(int)in[0]] << 18
+            long tmp = REVERSE_ALPHABET[(int)in[0]] << 18
                     | REVERSE_ALPHABET[(int)in[1]] << 12
                     | REVERSE_ALPHABET[(int)in[2]] << 6;
             out[0] = tmp >> 16 & 255;
